+++ lib/bwapi/client/brandwatch/clients/command_center/users/access.rb
@@ -10,14 +10,14 @@ module BWAPI
               #
               # TODO: Add parameters documentation
               def brandwatch_update_user_access_level(client_id, user_id, access_level)
-                put "brandwatch/client/#{client_id}/commandcenter/users/#{user_id}/access/#{access_level}"
+                put "brandwatch/clients/#{client_id}/commandcenter/users/#{user_id}/access/#{access_level}"
               end
 
               # Revoke a user's Vizia access
               #
               # TODO: Add parameters documentation
               def brandwatch_delete_user_access_level(client_id, user_id, access_level)
-                delete "brandwatch/client/#{client_id}/commandcenter/users/#{user_id}/access/#{access_level}"
+                delete "brandwatch/clients/#{client_id}/commandcenter/users/#{user_id}/access/#{access_level}"
               end
             end
           end