+++ lib/bwapi/client/client.rb
@@ -1,3 +1,7 @@
+require 'bwapi/client/client/grant_types'
+require 'bwapi/client/client/pricing_options'
+require 'bwapi/client/client/sub_clients'
+
 module BWAPI
   class Client
     # Client module for client endpoints
@@ -9,12 +13,37 @@ module BWAPI
         get 'client'
       end
 
+      # Get the current users client subclients
+      #
+      # @return [Hash] All sub clients information for user's client
+      def client_subclients
+        get 'client/subclients'
+      end
+
       # Get the current users client modules
       #
       # @return [Hash] Client module information
       def client_modules
         get 'client/modules'
       end
+
+      # Given a project id, retrieves the price structure associated with the client the project belongs to
+      #
+      # TODO: Add parameters documentation
+      def client_pricing(opts = {})
+        get '/client/pricing', opts
+      end
+
+      # Fetch usage report
+      #
+      # TODO: Add parameters documentation
+      def client_usage
+        get '/client/usage'
+      end
+
+      include BWAPI::Client::Client::GrantTypes
+      include BWAPI::Client::Client::PricingOptions
+      include BWAPI::Client::Client::SubClients
     end
   end
 end