+++ spec/bwapi/default_spec.rb
@@ -32,9 +32,9 @@ describe BWAPI::Default do
         api_endpoint: 'https://newapi.brandwatch.com/',
         client_id: 'brandwatch-api-client',
         client_secret: nil,
-        connection_options:  {
-          headers:  {
-            user_agent: 'BWAPI Ruby Gem 11.0.0'
+        connection_options: {
+          headers: {
+            user_agent: 'BWAPI Ruby Gem 12.0.0'
           },
           request: {
             params_encoder: Faraday::FlatParamsEncoder
@@ -46,7 +46,7 @@ describe BWAPI::Default do
         password: nil,
         performance: {},
         refresh_token: nil,
-        user_agent: 'BWAPI Ruby Gem 11.0.0',
+        user_agent: 'BWAPI Ruby Gem 12.0.0',
         username: nil,
         verify_ssl: false
       )
@@ -128,7 +128,7 @@ describe BWAPI::Default do
     it 'should return the default hash values' do
       expect(BWAPI::Default.connection_options).to eql(
         headers: {
-          user_agent: 'BWAPI Ruby Gem 11.0.0'
+          user_agent: 'BWAPI Ruby Gem 12.0.0'
         },
         request: {
           params_encoder: Faraday::FlatParamsEncoder